+++ src/main/java/lotto/Main.java
@@ -1,16 +1,17 @@
 package lotto;
 
+import lotto.controller.LottoController;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class Main {
 
     public static void main(String[] args) {
-        LottoController controller = initController();
-        controller.start();
+        LottoController lottoController = initializeController();
+        lottoController.start();
     }
 
-    private static LottoController initController() {
+    private static LottoController initializeController() {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
         return new LottoController(inputView, outputView);
