+++ src/main/java/lotto/model/DashBoard.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.model;
 
 import java.util.EnumMap;
+import lotto.Rank;
 
 public class DashBoard {
