+++ src/main/java/lotto/model/RankDeterminer.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.model;
+
+import lotto.Rank;
 
 public class RankDeterminer {
