+++ src/test/java/lotto/model/CashierTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
