+++ src/test/java/lotto/model/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
